dates.get_run_dates: Refactor to always use environment
Since we no longer generate any times and are just doing math on
dates, I don't believe we need the timezone information any more. It
looks like `datetime.datetime.combine` uses the timezone of the time
it's working with, which in this case is `datetime.time.min`, and
which defaults to `None`, according to the docs. That part hasn't been
changed, so the timezone of the results should be the same as before.

This also adds a test around the date generation, to give a little
more confidence that it works appropriately.