get_stock: Use correct date
This was accidentally inserting current data for whatever date was
being backfilled. Now it always puts yesterday's date on the stock
information. This is sorta correct, but could result in weird data if
run during the middle of a day. So we also check if the date being run
for is yesterday, which should mostly happen when running scheduled
runs. There is a risk that rerunning this could result in odd values
being inserted, but I don't think it's worth more effort right now.